Extract configuring `JacocoReport` into a script plugin
